Rename presentational comps to better suit purpose
- in imports feature module
